# backend/portfolio_app/static/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"description": "API del Portfolio de RuizDev7 con sistema de roles y permisos. Esta API proporciona endpoints para gestionar posts, bombas, usuarios, categorías, comentarios y el sistema de autorización basado en roles.",
		"version": "1.0.0",
		"title": "Portfolio RuizDev7 API",
		"license": {
			"name": "MIT",
			"url": "https://opensource.org/licenses/MIT"
		},
		"contact": {
			"name": "RuizDev7",
			"email": "ruizdev7@outlook.com"
		}
	},
	"servers": [
		{
			"url": "http://localhost:8000",
			"description": "Servidor de desarrollo"
		},
		{
			"url": "https://api.ruizdev7.com",
			"description": "Servidor de producción"
		}
	],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Token JWT obtenido del endpoint /api/v1/token"
			}
		},
		"schemas": {
			"User": {
				"type": "object",
				"properties": {
					"ccn_user": {
						"type": "integer",
						"example": 1,
						"description": "ID único del usuario"
					},
					"first_name": {
						"type": "string",
						"example": "Joseph",
						"description": "Primer nombre del usuario"
					},
					"middle_name": {
						"type": "string",
						"example": "G",
						"description": "Segundo nombre del usuario (opcional)"
					},
					"last_name": {
						"type": "string",
						"example": "Ruiz",
						"description": "Apellido del usuario"
					},
					"email": {
						"type": "string",
						"format": "email",
						"example": "ruizdev7@outlook.com",
						"description": "Email único del usuario"
					},
					"account_id": {
						"type": "string",
						"example": "hash_generado_automaticamente",
						"description": "ID de cuenta único generado automáticamente"
					},
					"created_at": {
						"type": "string",
						"format": "date-time",
						"example": "2024-01-15T10:30:00Z",
						"description": "Fecha de creación del usuario"
					}
				}
			},
...
			"LoginRequest": {
				"type": "object",
				"required": [
					"email",
					"password"
				],
				"properties": {
					"email": {
						"type": "string",
						"format": "email",
						"example": "ruizdev7@outlook.com",
						"description": "Email del usuario"
					},
					"password": {
						"type": "string",
						"example": "password123",
						"description": "Contraseña del usuario"
					}
				}
			},
			"LoginResponse": {
				"type": "object",
				"properties": {
					"current_user": {
						"type": "object",
						"properties": {
							"user_info": {
								"$ref": "#/components/schemas/User"
							},
							"token": {
								"type": "string",
								"example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...",
								"description": "Token JWT de acceso"
							},
							"refresh_token": {
								"type": "string",
								"example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...",
								"description": "Token JWT de refresco"
							},
							"account_id": {
								"type": "string",
								"example": "hash_value",
								"description": "ID de cuenta único"
							},
							"roles": {
								"type": "array",
								"items": {
									"type": "object",
									"properties": {
										"ccn_role": {
											"type": "integer",
											"example": 1
										},
										"role_name": {
											"type": "string",
											"example": "admin"
										}
									}
								}
							},
							"permissions": {
								"type": "array",
								"items": {
									"type": "object",
									"properties": {
										"permission_name": {
											"type": "string",
											"example": "posts_create"
										},
										"resource": {
											"type": "string",
											"example": "posts"
										},
										"action": {
											"type": "string",
											"example": "create"
										},
										"description": {
											"type": "string",
											"example": "Crear posts"
										}
									}
								}
							}
						}
					}
				}
			},
			"Error": {
				"type": "object",
				"properties": {
					"error": {
						"type": "string",
						"example": "Mensaje de error",
						"description": "Descripción del error"
					},
					"msg": {
						"type": "string",
						"example": "Mensaje de error",
						"description": "Mensaje de error alternativo"
					}
				}
			}
		}
	},
	"paths": {
		"/api/v1/token": {
			"post": {
				"tags": [
					"Autenticación"
				],
				"summary": "Iniciar sesión",
				"description": "Autentica un usuario y devuelve tokens JWT junto con información de roles y permisos",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Login exitoso",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginResponse"
								}
							}
						}
					},
					"401": {
						"description": "Credenciales inválidas",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			}
		},
		"/api/v1/refresh-token": {
			"get": {
				"tags": [
					"Autenticación"
				],
				"summary": "Refrescar token",
				"description": "Refresca el token de acceso usando el token de refresco",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "Token refrescado exitosamente",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"New access token": {
											"type": "string",
											"example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/v1/logout": {
			"delete": {
				"tags": [
					"Autenticación"
				],
				"summary": "Cerrar sesión",
				"description": "Revoca el token JWT actual",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "Logout exitoso",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"example": "access token revoked successfully"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/v1/whoami": {
			"get": {
				"tags": [
					"Autenticación"
				],
				"summary": "Información del usuario actual",
				"description": "Obtiene información del usuario autenticado incluyendo roles y permisos",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "Información del usuario",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"email": {
											"type": "string",
											"example": "ruizdev7@outlook.com"
										},
										"ccn_user": {
											"type": "integer",
											"example": 1
										},
										"first_name": {
											"type": "string",
											"example": "Joseph"
										},
										"last_name": {
											"type": "string",
											"example": "Ruiz"
										},
										"roles": {
											"type": "array",
											"items": {
												"type": "object",
												"properties": {
													"ccn_role": {
														"type": "integer",
														"example": 1
													},
													"role_name": {
														"type": "string",
														"example": "admin"
													}
												}
											}
										},
										"permissions": {
											"type": "array",
											"items": {
												"type": "object"
											}
										}
									}
								}
							}
						}
					}
				}
			}
		},
...
	},
...
}